Set LD_LIBRARY_PATH when running executables from builddir
Unlike libtool, which generates wrapper scripts for built programs to
run without installation, meson relies on RPATH to load shared libraries
from builddir. However, the system compiler on FreeBSD are configured to
pass --enable-new-dtags to the linker, so RUNPATH is used instead of
RPATH. If LD_LIBRARY_PATH is set in the environment, RUNPATH will be
overrided and libraries will be loaded from LD_LIBRARY_PATH instead of
the path specified by meson with -Wl,-rpath.

To workaround the meson bug, a simple shell script is added to set
LD_LIBRARY_PATH for running cd-create-profile and cd-it8 from
custom_target.

https://github.com/mesonbuild/meson/issues/2881